requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-chroma==0.1.0
streamlit==1.29.0
python-dotenv==1.0.0
pypdf==3.17.0
chromadb==0.4.22
openai==1.7.0
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
tiktoken==0.5.2
